Extract Sluggable concern to GobiertoCommon
