+++ app/models/concerns/gobierto_common/sluggable.rb
@@ -1,0 +1,26 @@
+module GobiertoCommon
+  module Sluggable
+    extend ActiveSupport::Concern
+
+    included do
+      before_create :set_slug
+    end
+
+    private
+
+    def set_slug
+      return if self.slug.present?
+
+      base_slug = attributes_for_slug.join('-').gsub('_', ' ').parameterize
+      new_slug  = base_slug
+
+      count = 1
+      while self.class.exists?(site: site, slug: new_slug)
+        new_slug = "#{base_slug}-#{count}"
+        count += 1
+      end
+
+      self.slug = new_slug
+    end
+  end
+end
