+++ test/models/gobierto_people/person_statement_test.rb
@@ -1,11 +1,11 @@
 require "test_helper"
 require "support/concerns/user/subscribable_test"
-require "support/concerns/gobierto_people/sluggable_test"
+require "support/concerns/gobierto_common/sluggable_test"
 
 module GobiertoPeople
   class PersonStatementTest < ActiveSupport::TestCase
     include User::SubscribableTest
-    include ::GobiertoPeople::SluggableTestModule
+    include GobiertoCommon::SluggableTestModule
 
     def person_statement
       @person_statement ||= gobierto_people_person_statements(:richard_current)
@@ -13,7 +13,7 @@ module GobiertoPeople
     alias subscribable person_statement
 
     def new_person_statement
-      ::GobiertoPeople::PersonStatement.create!(
+      GobiertoPeople::PersonStatement.create!(
         title: 'Person Statement Title',
         published_on: Time.zone.now,
         person: gobierto_people_people(:richard),
